add todo to delete in Order Controller

+++ src/main/java/com/orderandwarehouse/app/controller/OrderController.java
@@ -48,8 +48,15 @@ public class OrderController {
 
     @DeleteMapping("/{id}")
     public HttpStatus delete(@PathVariable Long id) {
-        service.delete(id);
-        return HttpStatus.OK;
+         /*
+        TODO check if ID exists? OK : NOT_FOUND
+                if ID exists check if it's in use? StorageUnitStillInUseException : OK
+        */
+        if (service.getById(id).isPresent()) {
+            service.delete(id);
+            return HttpStatus.OK;
+        }
+        return HttpStatus.NOT_FOUND;
     }
 
 }
